+++ forui/lib/src/widgets/time_field/picker/picker_time_field.dart
@@ -63,6 +63,7 @@ class _PickerTimeField extends FTimeField implements FTimeFieldPickerProperties 
     super.builder,
     super.prefixBuilder,
     super.suffixBuilder,
+    super.clearable,
     super.label,
     super.description,
     super.enabled = true,
@@ -96,6 +97,12 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
   final TextEditingController _textController = .new();
   late FocusNode _focus = widget.focusNode ?? .new(debugLabel: 'FTimeField');
   DateFormat? _format;
+
+  @override
+  void initState() {
+    super.initState();
+    _textController.addListener(_onTextChange);
+  }
 
   @override
   void didChangeDependencies() {
@@ -141,8 +148,16 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
     if (widget.focusNode == null) {
       _focus.dispose();
     }
-    _textController.dispose();
+    _textController
+      ..removeListener(_onTextChange)
+      ..dispose();
     super.dispose();
+  }
+
+  void _onTextChange() {
+    if (_textController.text.isEmpty) {
+      _controller.value = null;
+    }
   }
 
   @override
@@ -153,10 +168,13 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
 
   void _updateTextController() {
     SchedulerBinding.instance.addPostFrameCallback((_) {
-      if (_controller.picker.value case final value) {
-        final time = value.withDate(DateTime(1970));
-        _textController.text = widget.format?.format(time) ?? _format?.format(time) ?? '';
-      }
+      _textController.text = switch (_controller.value) {
+        null => '',
+        final value =>
+          widget.format?.format(value.withDate(DateTime(1970))) ??
+              _format?.format(value.withDate(DateTime(1970))) ??
+              '',
+      };
     });
   }
 
@@ -184,9 +202,10 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
         expands: widget.expands,
         mouseCursor: widget.mouseCursor,
         canRequestFocus: widget.canRequestFocus,
+        clearable: widget.clearable ? (value) => value.text.isNotEmpty : (_) => false,
         onTap: _onTap,
         onTapAlwaysCalled: true,
-        hint: widget.hint ?? localizations.dateFieldHint,
+        hint: widget.hint ?? localizations.timeFieldHint,
         readOnly: true,
         enableInteractiveSelection: false,
         prefixBuilder: widget.prefixBuilder,